Added non-reactive public interface
