+++ UnsplashKit/Classes/Clients/Source/UnsplashClient+Calls.swift
@@ -9,31 +9,39 @@ import RxSwift
 
 extension UnsplashClient {
 
-    public func randomPhoto(_ size: CGSize? = nil, filter: SourceRequestFilter? = .none) -> Observable<UnsplashImage> {
-        return self.request(request: self.requestFactory.random(size, filter: filter)).map { $0.0 }
+    public typealias ImageCompletion = (_ image: UnsplashImage) -> Void
+
+    public func randomPhoto(_ size: CGSize? = nil, filter: SourceRequestFilter? = .none, completion: @escaping ImageCompletion) {
+        process(observable: randomPhoto(size, filter: filter), forCompletion: completion)
     }
 
-    public func randomPhoto(fromSearch terms: [String], size: CGSize? = nil, filter: SourceRequestFilter? = .none) -> Observable<UnsplashImage> {
-        return self.request(request: self.requestFactory.search(terms, size: size, filter: filter)).map { $0.0 }
+    public func randomPhoto(fromSearch terms: [String], size: CGSize? = nil, filter: SourceRequestFilter? = .none, completion: @escaping ImageCompletion) {
+        process(observable: randomPhoto(fromSearch: terms, size: size, filter: filter), forCompletion: completion)
     }
 
-    public func randomPhoto(fromCategory category: SourceCategory, size: CGSize? = nil, filter: SourceRequestFilter? = .none) -> Observable<UnsplashImage> {
-        return self.request(request: self.requestFactory.category(category, size: size, filter: filter)).map { $0.0 }
+    public func randomPhoto(fromCategory category: SourceCategory, size: CGSize? = nil, filter: SourceRequestFilter? = .none, completion: @escaping ImageCompletion) {
+        process(observable: randomPhoto(fromCategory: category, size: size, filter: filter), forCompletion: completion)
     }
 
-    public func randomPhoto(fromUser userID: String, size: CGSize? = nil, filter: SourceRequestFilter? = .none) -> Observable<UnsplashImage> {
-        return self.request(request: self.requestFactory.user(userID, size: size, filter: filter)).map { $0.0 }
+    public func randomPhoto(fromUser userID: String, size: CGSize? = nil, filter: SourceRequestFilter? = .none, completion: @escaping ImageCompletion) {
+        process(observable: randomPhoto(fromUser: userID, size: size, filter: filter), forCompletion: completion)
     }
 
-    public func randomPhoto(fromUserLikes userID: String, size: CGSize? = nil) -> Observable<UnsplashImage> {
-        return self.request(request: self.requestFactory.userLikes(userID, size: size)).map { $0.0 }
+    public func randomPhoto(fromUserLikes userID: String, size: CGSize? = nil, completion: @escaping ImageCompletion) {
+        process(observable: randomPhoto(fromUserLikes: userID, size: size), forCompletion: completion)
     }
 
-    public func randomPhoto(fromCollection collectionID: String, size: CGSize? = nil) -> Observable<UnsplashImage> {
-        return self.request(request: self.requestFactory.collection(collectionID, size: size)).map { $0.0 }
+    public func randomPhoto(fromCollection collectionID: String, size: CGSize? = nil, completion: @escaping ImageCompletion) {
+        process(observable: randomPhoto(fromCollection: collectionID, size: size), forCompletion: completion)
     }
 
-    public func photo(_ identifier: String, size: CGSize? = nil) -> Observable<UnsplashImage> {
-        return self.request(request: self.requestFactory.photo(identifier, size: size)).map { $0.0 }
+    public func photo(_ identifier: String, size: CGSize? = nil, completion: @escaping ImageCompletion) {
+        process(observable: photo(identifier, size: size), forCompletion: completion)
+    }
+
+    private func process(observable: Observable<UnsplashImage>, forCompletion completion: @escaping ImageCompletion) {
+        observable.subscribe(onNext: { image in
+            completion(image)
+        })
     }
 }
